# Improve throughput of TimeSpan.ToString/TryFormat with "g"/"G"