TimeSpan has three standard formats: "c", "g", and "G".  Yesterday I updated its implementation with throughput improvements for "c" (the default) based on porting the design from Utf8Formatter; this PR does so for "g"/"G".

Initially I wasn't going to handle "g"/"G" as they factor in culture (Utf8Formatter doesn't), but even with accessing the current culture there are still significant wins to be had.  I was also going to keep the "c" and "g"/"G" implementations separate, to avoid bogging down the default "c" formatting with additional conditions needed to support "g"/"G", but the overhead incurred for that turns out to be minimal enough that it's worth keeping one implementation rather than two mostly-similar ones... the impact on "c" is mostly within noise.

This PR makes a significant throughput improvement for "g"/"G" formatting.  It also removes several unnecessary allocations, such that TryFormat with "g"/"G" is now allocation-free (and ToString just allocates the asked-for string).

Benchmark:
```C#
using System;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
[InProcess]
public class Benchmark
{
    private static void Main() => BenchmarkRunner.Run<Benchmark>();
    private char[] _tmp = new char[100];

    [Benchmark] public string ToStringMaxg() => TimeSpan.MaxValue.ToString("g");
    [Benchmark] public bool TryFormatMaxg() => TimeSpan.MaxValue.TryFormat(_tmp, out int charsWritten, "g");

    [Benchmark] public string ToStringMaxG() => TimeSpan.MaxValue.ToString("G");
    [Benchmark] public bool TryFormatMaxG() => TimeSpan.MaxValue.TryFormat(_tmp, out int charsWritten, "G");
}
```

Before:
```
        Method |      Mean |      Error |     StdDev |    Median |  Gen 0 | Allocated |
-------------- |----------:|-----------:|-----------:|----------:|-------:|----------:|
  ToStringMaxg | 687.27 ns | 13.8438 ns | 37.1906 ns | 679.25 ns | 0.1478 |     312 B |
 TryFormatMaxg | 677.25 ns | 16.8290 ns | 47.1903 ns | 670.88 ns | 0.1135 |     240 B |
  ToStringMaxG | 398.34 ns |  9.8501 ns |  9.6741 ns | 398.94 ns | 0.1292 |     272 B |
 TryFormatMaxG | 411.79 ns |  9.2157 ns | 10.9706 ns | 411.81 ns | 0.0949 |     200 B |
```

After:
```
        Method |     Mean |     Error |    StdDev |   Median |  Gen 0 | Allocated |
-------------- |---------:|----------:|----------:|---------:|-------:|----------:|
  ToStringMaxg | 79.73 ns | 1.0577 ns | 0.9377 ns | 79.72 ns | 0.0342 |      72 B |
 TryFormatMaxg | 67.84 ns | 0.8463 ns | 0.7067 ns | 67.76 ns |      - |       0 B |
  ToStringMaxG | 90.14 ns | 3.3078 ns | 3.8093 ns | 89.38 ns | 0.0342 |      72 B |
 TryFormatMaxG | 69.25 ns | 1.0194 ns | 0.9036 ns | 69.19 ns |      - |       0 B |
```

Contributes to https://github.com/dotnet/corefx/issues/30612
cc: @jkotas, @danmosemsft, @ahsonkhan