@stephentoub 2018-07-20
@dotnet-bot test Ubuntu x64 Checked Innerloop Build and Test please